Fold existing outdoor and indoor maps into one fragment
