+++ app/src/main/res/layout/fragment_maps.xml
@@ -101,14 +101,15 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/tabLayout" />
 
-    <androidx.viewpager.widget.ViewPager
+
+
+    <FrameLayout
         android:id="@+id/mapsContainer"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/horizLine">
 
-    </androidx.viewpager.widget.ViewPager>
+    </FrameLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
